python-package: disable albumentations version check (#2769)
When albumentations is imported, it checks its version and warns the
user if it is not running its latest version.  This is unnecessarily
noisy.  The goal of the warning is to prevent upstream from receiving
bug reports about issues on old versions.  Upstream recommends that
libraries that rely on it disable it themselves in this manner, see
https://github.com/albumentations-team/albumentations/issues/2206#issuecomment-2585905414